videogames.json añadido

+++ api/seeds/games.js
@@ -1,35 +1,39 @@
-// API DATOS --> MONGO
 const mongodb = require('../db/conn');
 const rawg = require('../services/rawg');
 const fs = require("fs");
 
 const seedGames = async () => {
-    // Añade los Juegos a mongo
-    // Conexión a Mongo
     await mongodb.connectToDatabase();
     const db = mongodb.getDb();
 
-    // Todos los games
     let totalGames = [];
 
-    // Hay varias páginas de juegos (Consultar Documentación API)
-    const games = await rawg.getGames();
-    totalGames = totalGames.concat(games);
-    console.log("Total games:",totalGames.length); // Comprobar
+    // Si el dataset ya existe usa los games del JSON
+    if (fs.existsSync("./datasets/videogames.json")) {
 
+        const data = fs.readFileSync("./datasets/videogames.json");
+        totalGames = JSON.parse(data);
     
-    await db.collection("videogames").insertMany(totalGames); //ESTA LINEA NO AÑADE A MONGO
-    console.log("RAWG TotalGames a Mongo");
+    // Si el dataset no existe: pasa los games de RAWG al JSON
+    } else {
 
+        const games = await rawg.getGames();
+        totalGames = totalGames.concat(games);
 
-    // ---------------
-    // Añade los juegos a ../datasets/videogames.json
-    fs.writeFileSync(
-        "./datasets/videogames.json",
-        JSON.stringify(totalGames, null, 2)
-    );
+        fs.writeFileSync(
+            "./datasets/videogames.json",
+            JSON.stringify(totalGames, null, 2)
+        );
 
-    console.log("videogames.json creado");
+        console.log("Dataset JSON creado");
+
+    }
+
+    //  JSON -> MongoDB
+    await db.collection("videogames").deleteMany({});
+    await db.collection("videogames").insertMany(totalGames);
+
+    console.log("Datos cargados en MongoDB");
 
     process.exit();
 };
